feat: refactoring - first iteration
