+++ src/domain/action-params.model.ts
@@ -1,0 +1,9 @@
+/**
+ * Action parameters
+ */
+export interface ActionParams {
+    branchPattern: string;
+    failIfInvalidBranchName: string;
+    ignoreBranchPattern: string;
+    commentForInvalidBranchName: string;
+}
